+++ core/src/main/java/air/ApplicationContext.java
@@ -1,0 +1,10 @@
+package air;
+
+public class ApplicationContext {
+
+    // 빈 생성
+    // 등록
+    public void initializeContext() {
+
+    }
+}
