+++ settings.gradle
@@ -3,3 +3,5 @@ rootProject.name = 'building-tomcat-spring'
 include 'mvc'
 include 'app'
 include 'learning'
+include 'core'
+
